♻️ refactor(eest): Migrate env_init to eest make config
