+++ docs/library/cli/index.md
@@ -2,4 +2,3 @@
 
 - [`eest`](eest.md) - A CLI tool that helps with routine tasks in ethereum/execution-spec-tests.
 - [`evm_bytes`](evm_bytes.md) - Convert the given EVM bytes from a binary file or a hex string to EEST's python opcodes.
-- [`env_init`](env_init.md) - Generate an example `env.yaml` [environment](../../dev/configurations.md) config file if it doesn't exist.
