+++ src/cli/eest/make/cli.py
@@ -10,6 +10,8 @@ is present, it shows a list of valid subcommands to choose from.
 
 import click
 
+from cli.eest.make.commands import create_default_config
+
 from .commands import test
 
 
@@ -32,3 +34,4 @@ Register nested commands here. For more information, see Click documentation:
 https://click.palletsprojects.com/en/8.0.x/commands/#nested-handling-and-contexts
 """
 make.add_command(test)
+make.add_command(create_default_config)
